Using try-with-resources and adding file encoding
We tried using the pt-BR locale before, which didn't change anyhting and
just created some problems with default encoding. Now we removed the
useless default encoding set and set UTF-8 as the encoding for the files
created by the program to avoid similar issues in the future. Also used
try-with-resources instead of manually closing the file (which was being
done wrongly before, in no finally block).