Adding the Parallel.forEach API
Adding the Parallel.forEach API which uses a cached thread pool to
iterate through a collection using a maximum of 4 threads (number can be
easily adjusted). It performs significantly better than the normal for
when the operation can be efficiently parallelized, like for when handling
files or network operations (which aren't bottlenecked by the processor).